Operators function added
This function is written to take the user's clicked input for an
operator and call the matching function. Function isn't tested and I'm
not sure it works just yet.

Another issue that needs to be addressed is assigning the inputs on the
calculator to 'a' and 'b' parameters of these operator functions.