Restructering function based on user click
The main function has now become dependent on which buttons the user
clicks. The querySelectorAll now targets all button elements and goes
through an if...else if block to assign values for the first number.

Still need to finish the function to accept an operator and the second
number to get a final number/result when "=" is clicked.