Add an operation div to the display row
By adding an operation div in the display row, I am hoping to be able to
use it to assign an operator value. The plan is to use this div in the
if...else if block of code to check if both the display and operator
divs have child elements. If they do, then proceed to getting the second
number to calculate.