Restructure javascript
Decided to change the main layout of the javascript. Rather than use the
same event listener to apply to all the buttons, group event listeners
by button class.

Currently there are three event listeners for number, operator and
buttonclear but still need to work out the details of each block.