Event listener for operator buttons
Currently the main issue is trying to get multiple digits assigned to
numberB. Having two elements referring to the display div doesn't seem
to be helping as I thought it would so now there is only one div.

The event listener for operator buttons is written with an if statement
saying, if "op !== '='" then keep letting numbers be assigned to the
display div, to be assigned to numberB, but it only takes one number.
However, the number does change depending on the last number button that
is clicked.