Event listener for number buttons
Working on numberB issue: only able to take one digit. There is an empty
array to push values to when a button is clicked. When console logging
numberB, all the numbers show up but as single numbers. Trying to use
reduce() doesn't seem to produce the result I'm looking for.