HTML and CSS restructure
Change the top row class to "top-row" instead of just a regular row so
that it can be a container for the operation div and display-previous
div (which both are their own separate ID classes). In addition, added
a display-current div on the row below and removed the extra three
buttons that served no purpose.

Display-current div is styled to look like the main display window. The
overall calculator layout is tighter and some height, width parameters
are adjusted or removed completely. CSS is also restructed with general
classes and types near the top and IDs as they show up in the HTML.