JS event listeners
Add event listeners for numbers, operators and clear button.

*Numbers will display in the displayCur div. Currently no digit limit is
set. Event listener is shorter and references a separate function that
appends the div.

*Operators display in the top displayOp div. Still need to find a way to
limit one operator at a time. Event listener is shorter and references a
separate function that appends the div. Needs to be able to assign div
from displayCur to displayPre.

*Clear button clears divs from displayOp, displayPre, and displayCur.