sendOP function within operator event listener
Event listener changed to be an if...else if block that will append the
operator to the display if there isn't one already. If a new operator
button is clicked, that one will be the updated operator on the display.
This also limits the operators to occur one at a time.