Top left display div
Calculator is able to take three numbers and three operators but can't
compute the expression. Operator and number button event listeners
changed to make this possible by checking if displayPre, displayCur
and displayOp had divs. If they did, the expression gets moved to the
displayTopLeft div.

Currently, the calculator doesn't do all the math just yet. The
displayCur div also needs to display the latest number selected or be
erased.

There was also a small issue where the calculator wouldn't work properly
at startup unless the clear button was clicked first. This was resolved
by correcting a selector in the sendNum function.