Operator sendOp function computes multiple numbers
Change an else...if section within the sendOp function that performs the
operation of two numbers when another operator is clicked. This enables
multiple operations at a time like 1 + 2 * 3 / 4 etc.

One area of improvement could be to create a function that performs what
the equals button event listener does because it was copied into the
sendOp else...if block. Creating this function to contain the code block
and recalling it does not work. VSCode says the function is void.