Remove div "op-plus" and sendNum()
Removed div with id "op-plus" from HTML along with the const
displayTopLeft in JS because it was serving no purpose. Also removed any
lines containing this reference and adjusted surrounding code to make
the calculator function properly.

The sendNum() function was also just integrated into the number button
event listener because it was just one line that sends the number to the
displayCur div.

Currently need numbers to be limited to the size of the display divs.