Scientific notation for long numbers
The displayPre div displays numbers in scientific notation if they are
too large. A function called sciNotation is applied to all solution
variables within the operator and equals event listeners. This function
measures the length of an operand and if it exceeds a length, it gets
put through the toExponential method, if it does not exceed the length
it is returned.